requirements.txt
numpy
pandas
scipy
tqdm
biopython
pyBigWig
pybedtools